Moved all cart manipulation to CartContext, CartContext is now everything cart
